refactor: extract benchmark review gate cli commands
